fix-12: UI+Tile+Tests — 14 задач (TIL-C01,CHR-C01,UI-H03/04/06,TIL-H01,UI-M03/04/06,CHR-H01,TIL-C02,UI-L02/03,TIL-L02) temperature accumulation, ServiceLocator, Qi safe cast, Inventory integration, namespace fix, DateTime→ticks, Camera null guard, divide-by-zero (2026-04-11)

+++ UnityProject/Assets/Scripts/Tests/IntegrationTestScenarios.cs
@@ -1,7 +1,7 @@
 // ============================================================================
 // IntegrationTestScenarios.cs — Интеграционные тесты
 // Cultivation World Simulator
-// Версия: 1.0
+// Версия: 1.1 — Fix-12: PlayerSaveData CurrentQi/MaxQi float→long
 // ============================================================================
 // Создан: 2026-03-31
 // Этап: 6 - Testing & Balance
@@ -652,8 +652,8 @@ namespace CultivationGame.Tests
         public string Name;
         public int CultivationLevel;
         public int CultivationSubLevel;
-        public float CurrentQi;
-        public float MaxQi;
+        public long CurrentQi; // FIX: float→long (Fix-01 cascade, Fix-12 test fix) (2026-04-12)
+        public long MaxQi;    // FIX: float→long (Fix-01 cascade, Fix-12 test fix) (2026-04-12)
         public string CurrentLocationId;
         public float PlayTime;
     }
